Rename 'individualCustomer' to 'customerView'
