+++ src/features/customerView/applicationsCard/ApplicationsCard.stories.tsx
