+++ src/features/customerView/applicationsCard/ApplicationsCard.test.jsx
