+++ src/features/customerView/applicationsCard/ApplicationsCard.tsx
