+++ src/features/customerView/applicationsCard/applicationsCard.module.scss
