+++ src/features/customerView/boatsCard/BoatCertificate.tsx
