+++ src/features/customerView/boatsCard/BoatCertificates.tsx
