+++ src/features/customerView/boatsCard/__fixtures__/mockData.ts
