+++ src/features/customerView/boatsCard/__tests__/boatsCardUtils.test.ts
