+++ src/features/customerView/boatsCard/boatsCard.module.scss
