+++ src/features/customerView/boatsCard/boatsCardUtils.ts
