+++ src/features/customerView/individualCustomerPage.module.scss
