+++ src/features/customerView/leasesCard/leasesCard.module.scss
